+++ components/index.tsx
@@ -2,10 +2,12 @@ export { default as Button } from './button/Button'
 
 export { default as Modal } from './modal/Modal'
 
-export { default as Input } from './input/input'
+export { default as Input } from './input/Input'
 
 export { default as Alert } from './alert/alert'
 
 export { default as Message } from './message/message'
 
 export { Utils } from './common/utils'
+
+export { keycap } from './axios/index'